requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP client for API calls
requests==2.31.0
httpx==0.25.2

# OpenAI integration
openai==1.3.7

# Data processing
pandas==2.1.4
numpy==1.24.3

# Configuration management
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# File handling
pathlib2==2.3.7

# JSON processing
jsonschema==4.20.0

# Async support
asyncio-mqtt==0.16.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8 